Add _shipsLeft and overlap checks to Player.js
occupiedPoints is used as a global variable to check if a point that a user wants to add is already occupied by another ship.
I've made it an object with a push method since it acts as a "data structure" with checks.
There is some data duplication with Player.getOwnedShips()'s points iteration, that is to be removed in a future commit.
The currentShipPoints is a local array that is only relevant for a single ship, while the _occupiedPoints servers as a cache or a total of all points, occupied by a single player.

Add _shipsLeft checks to Player.js
- Add checks and tests to _shipsLeft
- Fix Game.js tests to not exceed ship limits